Use SIMD optimized sum for stochastic swap cost computation
As part of #7702 we added a function for performing a sum over an
array/Vec of f64 values that is optimized to let the compiler use
SIMD instructions to speed up the computation. The stochastic swap rust
module performs a very similar sum multiple times as part of each swap
trial. This commit moves the fast_sum() function used in the pauli
expectation value functions to a common location and updates stochastic
swap to use it too.
