+++ src/lib.rs
@@ -10,6 +10,7 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
+use std::convert::TryInto;
 use std::env;
 
 use pyo3::prelude::*;
@@ -33,6 +34,32 @@ pub fn getenv_use_multiple_threads() -> bool {
         .to_uppercase()
         == "TRUE";
     !parallel_context || force_threads
+}
+
+const LANES: usize = 8;
+
+// Based on the sum implementation in:
+// https://stackoverflow.com/a/67191480/14033130
+// and adjust for f64 usage
+#[inline]
+pub fn fast_sum(values: &[f64]) -> f64 {
+    let chunks = values.chunks_exact(LANES);
+    let remainder = chunks.remainder();
+
+    let sum = chunks.fold([0.; LANES], |mut acc, chunk| {
+        let chunk: [f64; LANES] = chunk.try_into().unwrap();
+        for i in 0..LANES {
+            acc[i] += chunk[i];
+        }
+        acc
+    });
+    let remainder: f64 = remainder.iter().copied().sum();
+
+    let mut reduced = 0.;
+    for val in sum {
+        reduced += val;
+    }
+    reduced + remainder
 }
 
 #[pymodule]
