+++ src/pauli_exp_val.rs
@@ -10,8 +10,6 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
-use std::convert::TryInto;
-
 use num_complex::Complex64;
 use numpy::PyReadonlyArray1;
 use pyo3::exceptions::PyOverflowError;
@@ -19,34 +17,9 @@ use pyo3::prelude::*;
 use pyo3::wrap_pyfunction;
 use rayon::prelude::*;
 
-use crate::getenv_use_multiple_threads;
+use crate::{fast_sum, getenv_use_multiple_threads};
 
-const LANES: usize = 8;
 const PARALLEL_THRESHOLD: usize = 19;
-
-// Based on the sum implementation in:
-// https://stackoverflow.com/a/67191480/14033130
-// and adjust for f64 usage
-#[inline]
-fn fast_sum(values: &[f64]) -> f64 {
-    let chunks = values.chunks_exact(LANES);
-    let remainder = chunks.remainder();
-
-    let sum = chunks.fold([0.; LANES], |mut acc, chunk| {
-        let chunk: [f64; LANES] = chunk.try_into().unwrap();
-        for i in 0..LANES {
-            acc[i] += chunk[i];
-        }
-        acc
-    });
-    let remainder: f64 = remainder.iter().copied().sum();
-
-    let mut reduced = 0.;
-    for val in sum {
-        reduced += val;
-    }
-    reduced + remainder
-}
 
 /// Compute the pauli expectatation value of a statevector without x
 #[pyfunction]
